+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## [Unrelead]
 
+### Added
+
+- The `--superuser-pw` option has been added to allow setting the "postgres" user's password.
+
 ### Changed
 
 - `startup_timeout` and `probe_delay` builder method signatures brought in line with the rest, taking a `&mut` receiver.